pyproject.toml
[project]
name = "genai-session"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastembed>=0.7.1",
    "hf-xet>=1.1.4",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "llama-index>=0.12.42",
    "llama-index-llms-openai>=0.4.7",
    "llama-index-utils-workflow>=0.3.4",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.3",
    "tavily-python>=0.7.6",
]
